LayoutTests/fast/js/global-constructors-expected.txt

This test documents our set of global constructors we expose on the window object (FF and Opera don't expose them on the window, btw). This also checks to make sure than any constructor attribute we expose has the expected constructor type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Attr.toString() is '[object AttrConstructor]'
PASS Audio.toString() is '[object AudioConstructor]'
PASS BeforeLoadEvent.toString() is '[object BeforeLoadEventConstructor]'
PASS CDATASection.toString() is '[object CDATASectionConstructor]'
PASS CSSCharsetRule.toString() is '[object CSSCharsetRuleConstructor]'
PASS CSSFontFaceRule.toString() is '[object CSSFontFaceRuleConstructor]'
PASS CSSImportRule.toString() is '[object CSSImportRuleConstructor]'
PASS CSSMediaRule.toString() is '[object CSSMediaRuleConstructor]'
PASS CSSPageRule.toString() is '[object CSSPageRuleConstructor]'
PASS CSSPrimitiveValue.toString() is '[object CSSPrimitiveValueConstructor]'
PASS CSSRule.toString() is '[object CSSRuleConstructor]'
PASS CSSRuleList.toString() is '[object CSSRuleListConstructor]'
PASS CSSStyleDeclaration.toString() is '[object CSSStyleDeclarationConstructor]'
PASS CSSStyleRule.toString() is '[object CSSStyleRuleConstructor]'
PASS CSSStyleSheet.toString() is '[object CSSStyleSheetConstructor]'
PASS CSSValue.toString() is '[object CSSValueConstructor]'
PASS CSSValueList.toString() is '[object CSSValueListConstructor]'
PASS CSSVariablesDeclaration.toString() is '[object CSSVariablesDeclarationConstructor]'
PASS CSSVariablesRule.toString() is '[object CSSVariablesRuleConstructor]'
PASS CanvasRenderingContext2D.toString() is '[object CanvasRenderingContext2DConstructor]'
PASS CharacterData.toString() is '[object CharacterDataConstructor]'
PASS ClientRect.toString() is '[object ClientRectConstructor]'
PASS ClientRectList.toString() is '[object ClientRectListConstructor]'
PASS Clipboard.toString() is '[object ClipboardConstructor]'
PASS Comment.toString() is '[object CommentConstructor]'
PASS Counter.toString() is '[object CounterConstructor]'
PASS DOMException.toString() is '[object DOMExceptionConstructor]'
PASS DOMImplementation.toString() is '[object DOMImplementationConstructor]'
PASS DOMParser.toString() is '[object DOMParserConstructor]'
PASS Document.toString() is '[object DocumentConstructor]'
PASS DocumentFragment.toString() is '[object DocumentFragmentConstructor]'
PASS DocumentType.toString() is '[object DocumentTypeConstructor]'
PASS Element.toString() is '[object ElementConstructor]'
PASS Entity.toString() is '[object EntityConstructor]'
PASS EntityReference.toString() is '[object EntityReferenceConstructor]'
PASS Event.toString() is '[object EventConstructor]'
PASS EventException.toString() is '[object EventExceptionConstructor]'
PASS File.toString() is '[object FileConstructor]'
PASS FileList.toString() is '[object FileListConstructor]'
PASS HTMLAllCollection.toString() is '[object HTMLAllCollectionConstructor]'
PASS HTMLAnchorElement.toString() is '[object HTMLAnchorElementConstructor]'
PASS HTMLAppletElement.toString() is '[object HTMLAppletElementConstructor]'
PASS HTMLAreaElement.toString() is '[object HTMLAreaElementConstructor]'
PASS HTMLAudioElement.toString() is '[object HTMLAudioElementConstructor]'
PASS HTMLBRElement.toString() is '[object HTMLBRElementConstructor]'
PASS HTMLBaseElement.toString() is '[object HTMLBaseElementConstructor]'
PASS HTMLBaseFontElement.toString() is '[object HTMLBaseFontElementConstructor]'
PASS HTMLBlockquoteElement.toString() is '[object HTMLBlockquoteElementConstructor]'
PASS HTMLBodyElement.toString() is '[object HTMLBodyElementConstructor]'
PASS HTMLButtonElement.toString() is '[object HTMLButtonElementConstructor]'
PASS HTMLCanvasElement.toString() is '[object HTMLCanvasElementConstructor]'
PASS HTMLCollection.toString() is '[object HTMLCollectionConstructor]'
PASS HTMLDListElement.toString() is '[object HTMLDListElementConstructor]'
PASS HTMLDataGridCellElement.toString() is '[object HTMLDataGridCellElementConstructor]'
PASS HTMLDataGridColElement.toString() is '[object HTMLDataGridColElementConstructor]'
PASS HTMLDataGridElement.toString() is '[object HTMLDataGridElementConstructor]'
PASS HTMLDirectoryElement.toString() is '[object HTMLDirectoryElementConstructor]'
PASS HTMLDivElement.toString() is '[object HTMLDivElementConstructor]'
PASS HTMLDocument.toString() is '[object HTMLDocumentConstructor]'
PASS HTMLElement.toString() is '[object HTMLElementConstructor]'
PASS HTMLEmbedElement.toString() is '[object HTMLEmbedElementConstructor]'
PASS HTMLFieldSetElement.toString() is '[object HTMLFieldSetElementConstructor]'
PASS HTMLFontElement.toString() is '[object HTMLFontElementConstructor]'
PASS HTMLFormElement.toString() is '[object HTMLFormElementConstructor]'
PASS HTMLFrameElement.toString() is '[object HTMLFrameElementConstructor]'
PASS HTMLFrameSetElement.toString() is '[object HTMLFrameSetElementConstructor]'
PASS HTMLHRElement.toString() is '[object HTMLHRElementConstructor]'
PASS HTMLHeadElement.toString() is '[object HTMLHeadElementConstructor]'
PASS HTMLHeadingElement.toString() is '[object HTMLHeadingElementConstructor]'
PASS HTMLHtmlElement.toString() is '[object HTMLHtmlElementConstructor]'
PASS HTMLIFrameElement.toString() is '[object HTMLIFrameElementConstructor]'
PASS HTMLImageElement.toString() is '[object HTMLImageElementConstructor]'
PASS HTMLInputElement.toString() is '[object HTMLInputElementConstructor]'
PASS HTMLIsIndexElement.toString() is '[object HTMLIsIndexElementConstructor]'
PASS HTMLLIElement.toString() is '[object HTMLLIElementConstructor]'
PASS HTMLLabelElement.toString() is '[object HTMLLabelElementConstructor]'
PASS HTMLLegendElement.toString() is '[object HTMLLegendElementConstructor]'
PASS HTMLLinkElement.toString() is '[object HTMLLinkElementConstructor]'
PASS HTMLMapElement.toString() is '[object HTMLMapElementConstructor]'
PASS HTMLMarqueeElement.toString() is '[object HTMLMarqueeElementConstructor]'
PASS HTMLMediaElement.toString() is '[object HTMLMediaElementConstructor]'
PASS HTMLMenuElement.toString() is '[object HTMLMenuElementConstructor]'
PASS HTMLMetaElement.toString() is '[object HTMLMetaElementConstructor]'
PASS HTMLModElement.toString() is '[object HTMLModElementConstructor]'
PASS HTMLOListElement.toString() is '[object HTMLOListElementConstructor]'
PASS HTMLObjectElement.toString() is '[object HTMLObjectElementConstructor]'
PASS HTMLOptGroupElement.toString() is '[object HTMLOptGroupElementConstructor]'
PASS HTMLOptionElement.toString() is '[object HTMLOptionElementConstructor]'
PASS HTMLParagraphElement.toString() is '[object HTMLParagraphElementConstructor]'
PASS HTMLParamElement.toString() is '[object HTMLParamElementConstructor]'
PASS HTMLPreElement.toString() is '[object HTMLPreElementConstructor]'
PASS HTMLQuoteElement.toString() is '[object HTMLQuoteElementConstructor]'
PASS HTMLScriptElement.toString() is '[object HTMLScriptElementConstructor]'
PASS HTMLSelectElement.toString() is '[object HTMLSelectElementConstructor]'
PASS HTMLStyleElement.toString() is '[object HTMLStyleElementConstructor]'
PASS HTMLTableCaptionElement.toString() is '[object HTMLTableCaptionElementConstructor]'
PASS HTMLTableCellElement.toString() is '[object HTMLTableCellElementConstructor]'
PASS HTMLTableColElement.toString() is '[object HTMLTableColElementConstructor]'
PASS HTMLTableElement.toString() is '[object HTMLTableElementConstructor]'
PASS HTMLTableRowElement.toString() is '[object HTMLTableRowElementConstructor]'
PASS HTMLTableSectionElement.toString() is '[object HTMLTableSectionElementConstructor]'
PASS HTMLTextAreaElement.toString() is '[object HTMLTextAreaElementConstructor]'
PASS HTMLTitleElement.toString() is '[object HTMLTitleElementConstructor]'
PASS HTMLUListElement.toString() is '[object HTMLUListElementConstructor]'
PASS HTMLVideoElement.toString() is '[object HTMLVideoElementConstructor]'
PASS Image.toString() is '[object ImageConstructor]'
PASS KeyboardEvent.toString() is '[object KeyboardEventConstructor]'
PASS MediaError.toString() is '[object MediaErrorConstructor]'
PASS MediaList.toString() is '[object MediaListConstructor]'
PASS MessageChannel.toString() is '[object MessageChannelConstructor]'
PASS MessageEvent.toString() is '[object MessageEventConstructor]'
PASS MessagePort.toString() is '[object MessagePortConstructor]'
PASS MimeType.toString() is '[object MimeTypeConstructor]'
PASS MimeTypeArray.toString() is '[object MimeTypeArrayConstructor]'
PASS MouseEvent.toString() is '[object MouseEventConstructor]'
PASS MutationEvent.toString() is '[object MutationEventConstructor]'
PASS NamedNodeMap.toString() is '[object NamedNodeMapConstructor]'
PASS Node.toString() is '[object NodeConstructor]'
PASS NodeFilter.toString() is '[object NodeFilterConstructor]'
PASS NodeList.toString() is '[object NodeListConstructor]'
PASS Notation.toString() is '[object NotationConstructor]'
PASS Option.toString() is '[object OptionConstructor]'
PASS OverflowEvent.toString() is '[object OverflowEventConstructor]'
PASS PageTransitionEvent.toString() is '[object PageTransitionEventConstructor]'
PASS Plugin.toString() is '[object PluginConstructor]'
PASS PluginArray.toString() is '[object PluginArrayConstructor]'
PASS ProcessingInstruction.toString() is '[object ProcessingInstructionConstructor]'
PASS ProgressEvent.toString() is '[object ProgressEventConstructor]'
PASS RGBColor.toString() is '[object RGBColorConstructor]'
PASS Range.toString() is '[object RangeConstructor]'
PASS RangeException.toString() is '[object RangeExceptionConstructor]'
PASS Rect.toString() is '[object RectConstructor]'
PASS SVGAngle.toString() is '[object SVGAngleConstructor]'
PASS SVGColor.toString() is '[object SVGColorConstructor]'
PASS SVGException.toString() is '[object SVGExceptionConstructor]'
PASS SVGGradientElement.toString() is '[object SVGGradientElementConstructor]'
PASS SVGLength.toString() is '[object SVGLengthConstructor]'
PASS SVGMarkerElement.toString() is '[object SVGMarkerElementConstructor]'
PASS SVGPaint.toString() is '[object SVGPaintConstructor]'
PASS SVGPathSeg.toString() is '[object SVGPathSegConstructor]'
PASS SVGPreserveAspectRatio.toString() is '[object SVGPreserveAspectRatioConstructor]'
PASS SVGRenderingIntent.toString() is '[object SVGRenderingIntentConstructor]'
PASS SVGTextContentElement.toString() is '[object SVGTextContentElementConstructor]'
PASS SVGTextPathElement.toString() is '[object SVGTextPathElementConstructor]'
PASS SVGTransform.toString() is '[object SVGTransformConstructor]'
PASS SVGUnitTypes.toString() is '[object SVGUnitTypesConstructor]'
PASS SharedWorker.toString() is '[object SharedWorkerConstructor]'
PASS Storage.toString() is '[object StorageConstructor]'
PASS StorageEvent.toString() is '[object StorageEventConstructor]'
PASS StyleSheet.toString() is '[object StyleSheetConstructor]'
PASS StyleSheetList.toString() is '[object StyleSheetListConstructor]'
PASS Text.toString() is '[object TextConstructor]'
PASS TextEvent.toString() is '[object TextEventConstructor]'
PASS TextMetrics.toString() is '[object TextMetricsConstructor]'
PASS UIEvent.toString() is '[object UIEventConstructor]'
PASS WebKitAnimationEvent.toString() is '[object WebKitAnimationEventConstructor]'
PASS WebKitCSSKeyframeRule.toString() is '[object WebKitCSSKeyframeRuleConstructor]'
PASS WebKitCSSKeyframesRule.toString() is '[object WebKitCSSKeyframesRuleConstructor]'
PASS WebKitCSSMatrix.toString() is '[object WebKitCSSMatrixConstructor]'
PASS WebKitCSSTransformValue.toString() is '[object WebKitCSSTransformValueConstructor]'
PASS WebKitPoint.toString() is '[object WebKitPointConstructor]'
PASS WebKitTransitionEvent.toString() is '[object WebKitTransitionEventConstructor]'
PASS WebSocket.toString() is '[object WebSocketConstructor]'
PASS WheelEvent.toString() is '[object WheelEventConstructor]'
PASS Worker.toString() is '[object WorkerConstructor]'
PASS XMLDocument.toString() is '[object DocumentConstructor]'
PASS XMLHttpRequest.toString() is '[object XMLHttpRequestConstructor]'
PASS XMLHttpRequestException.toString() is '[object XMLHttpRequestExceptionConstructor]'
PASS XMLHttpRequestUpload.toString() is '[object XMLHttpRequestUploadConstructor]'
PASS XMLSerializer.toString() is '[object XMLSerializerConstructor]'
PASS XPathEvaluator.toString() is '[object XPathEvaluatorConstructor]'
PASS XPathException.toString() is '[object XPathExceptionConstructor]'
PASS XPathResult.toString() is '[object XPathResultConstructor]'
PASS successfullyParsed is true

TEST COMPLETE
